requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
exceptiongroup==1.1.3
fake-useragent==1.3.0
h11==0.14.0
idna==3.4
ndg-httpsclient==0.5.1
numpy==1.26.1
outcome==1.3.0.post0
pandas==2.1.2
pyasn1==0.5.0
pycparser==2.21
pyOpenSSL==23.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.7
wsproto==1.2.0
